Parameter conversion type changed to float from int

+++ programming_paradigm/robust_division_calculator.py
@@ -1,8 +1,8 @@
 def safe_divide(numerator: int, denominator:int):
     
     try:
-        numerator = int(numerator)
-        denominator = int(denominator)
+        numerator = float(numerator)
+        denominator = float(denominator)
         result = numerator/denominator
         return f'The result of the division is {result:.1f}'
     except ZeroDivisionError:
